unsigned int -> uint*_t

+++ src/bitarr.c
@@ -1,7 +1,8 @@
 #include "bitarr.h"
 #include "bitops.h"
+#include <stdint.h>
 
-BitArray* BitArray_calloc(unsigned int n, unsigned int l)
+BitArray* BitArray_calloc(uint32_t n, uint8_t l)
 {
     unsigned int n_entries = 1 + (((l * n) - 1) / WORD_SIZE);
     // space for bitarray + space needed for n_entries of word_size
@@ -23,7 +24,7 @@ void BitArray_free(BitArray *bitarr) {
 }
 
 
-BitArray* BitArray_init(unsigned int A[], unsigned int n, unsigned int l)
+BitArray* BitArray_init(unsigned int A[], uint32_t n, uint8_t l)
 {
     BitArray* bit_arr = BitArray_calloc(n, l);
     // Compress values from A into BitArray
